+++ src/main/java/http/request/headers/Headers.java
@@ -1,4 +1,4 @@
-package http.headers;
+package http.request.headers;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
