+++ src/main/java/http/request/requestline/old/ProtocolAndVersion.java
@@ -1,4 +1,4 @@
-package http.requestline.old;
+package http.request.requestline.old;
 
 import java.util.Objects;
