+++ src/main/java/http/request/requestline/old/QueryStrings.java
@@ -1,4 +1,4 @@
-package http.requestline.old;
+package http.request.requestline.old;
 
 import java.util.Collections;
 import java.util.HashMap;
