+++ src/main/java/http/request/requestline/requestLine2/RequestLine2.java
@@ -1,4 +1,4 @@
-package http.requestline.requestLine2;
+package http.request.requestline.requestLine2;
 
 import java.util.Collections;
 import java.util.Map;
