+++ src/main/java/webserver/RequestHandler.java
@@ -5,9 +5,9 @@ import com.github.jknack.handlebars.Template;
 import com.github.jknack.handlebars.io.ClassPathTemplateLoader;
 import com.github.jknack.handlebars.io.TemplateLoader;
 import db.DataBase;
-import http.requestline.old.QueryStrings;
-import http.headers.Headers;
-import http.requestline.old.RequestLine;
+import http.request.requestline.old.QueryStrings;
+import http.request.headers.Headers;
+import http.request.requestline.old.RequestLine;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
