+++ src/test/java/http/RequestUtilsTest.java
@@ -1,8 +1,9 @@
 package http;
 
-import http.body.Body;
-import http.headers.Headers2;
-import http.requestline.requestLine2.RequestLine2;
+import http.request.RequestUtils;
+import http.request.body.Body;
+import http.request.headers.Headers2;
+import http.request.requestline.requestLine2.RequestLine2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
