+++ src/test/java/http/old/QueryStringsTest.java
@@ -1,6 +1,6 @@
 package http.old;
 
-import http.requestline.old.QueryStrings;
+import http.request.requestline.old.QueryStrings;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
