+++ src/test/java/http/requestLine/Path2Test.java
@@ -1,6 +1,6 @@
 package http.requestLine;
 
-import http.requestline.requestLine2.Path2;
+import http.request.requestline.requestLine2.Path2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
