+++ src/test/java/http/requestLine/Protocol2Test.java
@@ -1,6 +1,6 @@
 package http.requestLine;
 
-import http.requestline.requestLine2.Protocol2;
+import http.request.requestline.requestLine2.Protocol2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
